+++ test/utils/gameStateUtils.ts
@@ -11,8 +11,8 @@ import type {
 	Weapon,
 } from '#shared/types/index'
 import { buildFriendlyName, DEFAULT_ONION_UNIT_TYPE_ID, getUnitTypeCatalog, getWeaponType } from '#shared/unitDefinitions'
-import { buildBattlefieldDefenderView, buildBattlefieldOnionView } from '#web/lib/appViewHelpers'
-import { type BattlefieldDefenderView, type BattlefieldOnionView } from '#web/lib/battlefieldView'
+import { buildBattlefieldDefenderView, buildBattlefieldOnionView } from '#web/lib/battlefieldViewBuilders'
+import { getBattlefieldPosition, type BattlefieldOnionView, type BattlefieldUnit } from '#web/lib/battlefieldView'
 import { createGameClient, type GameClient, type GameSnapshot, type ScenarioMapSnapshot } from '#web/lib/gameClient'
 
 /**
@@ -97,34 +97,17 @@ export function makeOnion(overrides: Partial<OnionUnit> = {}): OnionUnit {
 
 export function makeBattlefieldDefender(
 	unitOverrides: Partial<DefenderUnit> = {},
-	viewOverrides: Partial<BattlefieldDefenderView> = {},
-): BattlefieldDefenderView {
+	viewOverrides: Partial<BattlefieldUnit> = {},
+): BattlefieldUnit {
 	const defender = makeDefender(unitOverrides)
-	const legacyView = viewOverrides as Partial<{
-		id: string
-		type: string
-		status: DefenderUnit['state']
-		move: number
-		squads: number
-		weaponDetails: ReadonlyArray<Weapon>
-	}>
-	const canonicalWeapons = Array.isArray(legacyView.weaponDetails)
-		? legacyView.weaponDetails
-		: Array.isArray(viewOverrides.weapons) ? viewOverrides.weapons : defender.weapons
 
 	return {
 		...buildBattlefieldDefenderView(defender, {
 			activePhase: 'DEFENDER_COMBAT',
 			activeTurnActive: true,
-			stackSize: viewOverrides.stackSize ?? legacyView.squads ?? 1,
+			stackSize: viewOverrides.squads ?? 1,
 		}),
 		...viewOverrides,
-		unitId: viewOverrides.unitId ?? legacyView.id ?? defender.unitId,
-		typeId: viewOverrides.typeId ?? legacyView.type ?? defender.typeId,
-		state: viewOverrides.state ?? legacyView.status ?? defender.state,
-		weapons: canonicalWeapons,
-		movesRemaining: viewOverrides.movesRemaining ?? legacyView.move ?? 0,
-		stackSize: viewOverrides.stackSize ?? legacyView.squads ?? 1,
 	}
 }
 
@@ -138,13 +121,10 @@ export function makeBattlefieldOnion(
 	}
 }
 
-export type BattlefieldDefenderFixture = Omit<Partial<BattlefieldDefenderView>, 'weapons'> & {
+export type BattlefieldDefenderFixture = Omit<Partial<BattlefieldUnit>, 'weaponDetails' | 'weapons'> & {
 	id: string
 	type: string
-	position: { q: number; r: number }
-	status: DefenderUnit['state']
-	move?: number
-	attack?: string
+	status: BattlefieldUnit['status']
 	weapons?: ReadonlyArray<Weapon> | string
 	weaponDetails?: ReadonlyArray<Partial<Weapon> & {
 		name?: string
@@ -156,7 +136,7 @@ export type BattlefieldDefenderFixture = Omit<Partial<BattlefieldDefenderView>, 
 	}>
 }
 
-export function canonicalizeBattlefieldDefenders(defenders: ReadonlyArray<BattlefieldDefenderFixture>): BattlefieldDefenderView[] {
+export function canonicalizeBattlefieldDefenders(defenders: ReadonlyArray<BattlefieldDefenderFixture>): BattlefieldUnit[] {
 	return defenders.map((view) => {
 		const { weaponDetails: legacyWeaponDetails, ...viewWithoutWeaponDetails } = view
 		const weaponDetails = legacyWeaponDetails?.map((weapon) => {
@@ -170,57 +150,39 @@ export function canonicalizeBattlefieldDefenders(defenders: ReadonlyArray<Battle
 			})
 		})
 
-		const fallbackWeaponTypeId = getUnitTypeCatalog()[view.type]?.weapons[0]?.typeId ?? `${view.type}.rifle`
-		const canonicalFallbackWeapons = typeof view.weapons === 'string' && weaponDetails === undefined
-			? [makeWeapon({ typeId: fallbackWeaponTypeId })]
-			: undefined
-
-		const canonicalViewOverrides: Partial<BattlefieldDefenderView> = {
-			movesRemaining: view.movesRemaining ?? view.move ?? 0,
-			stackSize: view.stackSize ?? (view as BattlefieldDefenderFixture & { squads?: number }).squads ?? 1,
-			actionableModes: view.actionableModes ?? [],
-			...(weaponDetails === undefined
-				? canonicalFallbackWeapons === undefined ? {} : { weapons: canonicalFallbackWeapons }
-				: { weapons: weaponDetails }),
-		}
-
 		return makeBattlefieldDefender({
 			unitId: view.unitId ?? view.id,
 			typeId: view.typeId ?? view.type,
-			position: view.position,
+			position: getBattlefieldPosition(view),
 			state: view.state ?? view.status,
 			friendlyName: view.friendlyName,
-		}, canonicalViewOverrides)
+			weapons: [],
+		}, {
+			...viewWithoutWeaponDetails,
+			...(weaponDetails === undefined ? {} : { weaponDetails }),
+		})
 	})
 }
 
 export function canonicalizeBattlefieldOnion(view: BattlefieldOnionView): BattlefieldOnionView {
-	const legacyView = view as unknown as {
-		id: string
-		type: string
-		status: OnionUnit['state']
-		rams: number
-		weaponDetails?: ReadonlyArray<Partial<Weapon> & { status?: Weapon['state'] }>
-	}
-	const weapons = (legacyView.weaponDetails ?? []).map((weapon) => makeWeapon({
+	const viewWithoutWeaponDetails = { ...view }
+	delete viewWithoutWeaponDetails.weaponDetails
+	const weapons = (view.weaponDetails ?? []).map((weapon) => makeWeapon({
 		...weapon,
-		typeId: weapon.typeId ?? `${legacyView.type}.${weapon.id?.replace(/-\d+$/, '')}`,
+		typeId: weapon.typeId ?? `${view.type}.${weapon.id.replace(/-\d+$/, '')}`,
 		state: weapon.state ?? (weapon as Weapon & { status?: Weapon['state'] }).status ?? 'ready',
 	}))
 
 	return makeBattlefieldOnion({
-		unitId: legacyView.id,
-		typeId: legacyView.type,
-		position: view.position,
-		state: legacyView.status,
+		unitId: view.id,
+		typeId: view.type,
+		position: getBattlefieldPosition(view),
+		state: view.status,
 		friendlyName: view.friendlyName,
 		treads: view.treads,
-		ramsRemaining: legacyView.rams,
+		ramsRemaining: view.rams,
 		weapons,
-	}, {
-		movesAllowed: view.movesAllowed,
-		movesRemaining: view.movesRemaining,
-	})
+	}, viewWithoutWeaponDetails)
 }
 
 /**
