+++ test/web/lib/stackSelection.test.ts
@@ -1,0 +1,13 @@
+import { describe, expect, it } from 'vitest'
+import { resolveBattlefieldStacksExpandable } from '../../../web/lib/stackSelection'
+
+describe('stackSelection', () => {
+  it('allows active defenders to expand stacks during movement or combat', () => {
+    expect(resolveBattlefieldStacksExpandable({
+      activeRole: 'defender',
+      activeTurnActive: true,
+      isCombatPhase: true,
+      isMovementPhase: false,
+    })).toBe(true)
+  })
+})
