+++ web/components/BattlefieldRightRail.tsx
@@ -3,7 +3,8 @@ import { ConfirmationSurface } from './ConfirmationSurface'
 import { CombatTargetList } from './CombatTargetList'
 import { BattlefieldInspectorPanel } from './BattlefieldInspectorPanel'
 import { InactiveEventStream } from './InactiveEventStream'
-import { isWeaponSelectionId, resolveBattlefieldUnitName } from '../lib/appViewHelpers'
+import { resolveBattlefieldUnitName } from '../lib/battlefieldNaming'
+import { isWeaponSelectionId } from '../lib/selectionIds'
 import { buildRightRailCombatPanelViewModel } from '../lib/rightRailCombatPanel'
 import type { BattlefieldDefenderView, BattlefieldOnionView } from '../lib/battlefieldView'
 import type { TimelineEvent } from '../lib/battlefieldView'
