+++ web/lib/useBattlefieldInteractionState.ts
@@ -3,7 +3,8 @@ import { findMovePath, type MoveMapSnapshot } from '../../shared/movePlanner'
 import { getUnitMovementAllowance } from '../../shared/unitMovement'
 import type { GameAction, ServerGameSnapshot } from './gameClient'
 import type { GameSessionController } from './gameSessionTypes'
-import { getAuthoritativeOnion, isWeaponSelectionId, resolveBattlefieldStackSelectionIds, resolveSelectionOwnerUnitId } from './appViewHelpers'
+import { getAuthoritativeOnion, resolveBattlefieldStackSelectionIds } from './stackSelection'
+import { isWeaponSelectionId, resolveSelectionOwnerUnitId } from './selectionIds'
 import { buildMoveCommitAction } from './commitActionBuilders'
 import { clearRightRailStackSelection, selectRightRailStackMembers, toggleRightRailStackMemberSelection } from './rightRailSelection'
 import type { TurnPhase } from '../../shared/types/index'
